fix(portability): clean up unsafe uses of `long` for LLP64
Following the portability audit in #41, `long` is 32-bit on LLP64
(Windows 64-bit) but 64-bit on LP64, causing pointer truncation,
file-size truncation, and accidental narrowing through APIs typed as
`long`. This commit fixes every site flagged in the issue.

## Confirmed bugs

- WebServer.cpp / php_core_lib.cpp: ftell's `long` return wraps for
  >2 GiB template files on Win64. Capture into a 64-bit local with
  bounds check before narrowing. Won't bite a real user (templates
  are tiny) but fixes the type confusion at the source.
- TextClient.h: `unsigned long lFileSize` widened to `uint64`, with
  matching `%llu` printf format in TextClient.cpp. Fixes amulecmd's
  mis-display of any shared file > 4 GiB on Win64.

## InternalEvents API widening

- `CMuleInternalEvent::SetExtraLong` / `GetExtraLong` (carrying a
  `long m_value`) renamed to `SetExtraInt64` / `GetExtraInt64` with
  underlying type `int64_t`. The existing `HTTPDownload.cpp:312`
  caller already routes a `wxFileOffset` (potentially > 4 GiB)
  through it via an explicit `(long)` cast that silently truncated
  on Win64; the cast is now dropped.
- All 16 callsites updated to the new name across `amule.cpp`,
  `amule-remote-gui.cpp`, `AsyncDNS.cpp`, `HTTPDownload.cpp`. Drop
  the stale `(long)m_result` cast at HTTPDownload.cpp:417 too.

## Pointer-to-integer casts → uintptr_t

- FileArea.cpp: pointer arithmetic in the SIGBUS handler used
  `unsigned long`, which would truncate pointers on Win64. POSIX-only
  path today, but worth fixing.
- MuleDebug.cpp: same in the bfd backtrace `s_pie_base` subtraction.

## Process IDs → int

- `webserver_pid` (amule.h, TerminationProcessAmuleweb.h/.cpp,
  amule.cpp call sites) changed from `long` to `int`. POSIX pid_t
  is typically `int`, Windows DWORDs fit in `int`, and `int` avoids
  `pid_t`'s portability wart on MSVC. Format strings updated from
  `%ld` to `%d`; the wxExecute return is now narrowed explicitly.
- AppImageIntegration.cpp: drop the `static_cast<long>(getpid())`
  -- getpid returns pid_t (int on every aMule target), no cast
  needed; format updated to `%d`.

## PartFileConvert intermediate

- Switch `long l + ToLong + (unsigned) cast` to
  `unsigned long l + ToULong + uint32 cast`. File indices are small
  in practice; the new pattern makes the type intent explicit.

Refs #41.

Build verification: macOS `cmake --build` of amule + amuled +
amulecmd + amuleweb + amulegui completes warning- and error-free on
the patched tree.

+++ src/AppImageIntegration.cpp
@@ -297,8 +297,8 @@ void PromptAndInstall(wxWindow* parent)
 	// save partfiles, release the EC port, etc., before the new instance
 	// tries to grab the same locks.
 	const wxString relaunchCmd = wxString::Format(
-		wxT("sh -c 'while kill -0 %ld 2>/dev/null; do sleep 0.2; done; exec \"%s\"'"),
-		static_cast<long>(getpid()),
+		wxT("sh -c 'while kill -0 %d 2>/dev/null; do sleep 0.2; done; exec \"%s\"'"),
+		static_cast<int>(getpid()),
 		appimagePath);
 	wxExecute(relaunchCmd, wxEXEC_ASYNC | wxEXEC_MAKE_GROUP_LEADER);
